pyproject.toml
[project]
name = "finalagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "zakia2324", email = "160940055+zakia2324@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
finalagent = "finalagent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
